feat: improve automatic domain configuration
