+++ apps/builder/package.json
@@ -91,6 +91,7 @@
     "css-tree": "^2.3.1",
     "debug": "^4.3.7",
     "downshift": "^6.1.7",
+    "entrijs": "^0.0.11",
     "fast-deep-equal": "^3.1.3",
     "immer": "^10.1.1",
     "immerhin": "^0.10.0",
@@ -109,7 +110,6 @@
     "react-colorful": "^5.6.1",
     "react-dom": "18.3.0-canary-14898b6a9-20240318",
     "react-error-boundary": "^5.0.0",
-    "react-script-hook": "^1.7.2",
     "remix-auth": "^3.7.0",
     "remix-auth-form": "^1.5.0",
     "remix-auth-github": "^1.7.0",
